ablog/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
fake-useragent==0.1.11
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
instagramy==3.2
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.6.0
python-decouple==3.4
pytz==2020.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
